// examples-dj/workflow-example/packages/workflow-server/src/language-server/generated/ast.ts
export type Node = ActivityNode | Category | TaskNode;

export const Node = 'Node';

export function isNode(item: unknown): item is Node {
    return reflection.isInstance(item, Node);
}

export type NodeType = 'decision' | 'fork' | 'join' | 'merge';

export function isNodeType(item: unknown): item is NodeType {
    return item === 'decision' || item === 'fork' || item === 'join' || item === 'merge';
}

export type TaskType = 'automated' | 'manual';

export function isTaskType(item: unknown): item is TaskType {
    return item === 'manual' || item === 'automated';
}

export type Weight = 'high' | 'low' | 'medium';

export function isWeight(item: unknown): item is Weight {
    return item === 'low' || item === 'medium' || item === 'high';
}

export interface ActivityNode extends AstNode {
    readonly $container: Model;
    readonly $type: 'ActivityNode';
    __id: string
    name: string
    nodeType?: NodeType
}

export const ActivityNode = 'ActivityNode';

export function isActivityNode(item: unknown): item is ActivityNode {
    return reflection.isInstance(item, ActivityNode);
}

export interface Category extends AstNode {
    readonly $container: Model;
    readonly $type: 'Category';
    __id: string
    children?: Model
    label?: string
    name: string
}

export const Category = 'Category';

export function isCategory(item: unknown): item is Category {
    return reflection.isInstance(item, Category);
}

export interface Edge extends AstNode {
    readonly $type: 'Edge' | 'WeightedEdge';
    __id: string
    source: Reference<Node>
    target: Reference<Node>
}

export const Edge = 'Edge';

export function isEdge(item: unknown): item is Edge {
    return reflection.isInstance(item, Edge);
}

export interface Model extends AstNode {
    readonly $container: Category;
    readonly $type: 'Model';
    edges: Array<Edge>
    metaInfos: Array<MetaInfo>
    nodes: Array<Node>
}

export const Model = 'Model';
// ...
export interface TaskNode extends AstNode {
    readonly $container: Model;
    readonly $type: 'TaskNode';
    __id: string
    duration?: number
    label?: string
    name: string
    reference?: string
    taskType?: TaskType
}

export const TaskNode = 'TaskNode';

export function isTaskNode(item: unknown): item is TaskNode {
    return reflection.isInstance(item, TaskNode);
}

export interface WeightedEdge extends Edge {
    readonly $type: 'WeightedEdge';
    __id: string
    source: Reference<Node>
    target: Reference<Node>
    weight: Weight
}

export const WeightedEdge = 'WeightedEdge';